pyproject.toml
[project]
name = "marimolib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "FumiHubCNS", email = "fendo@cns.s.u-tokyo.ac.jp" }
]
requires-python = ">=3.10"
dependencies = [
    "json5>=0.12.1",
    "marimo>=0.17.7",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "plotly>=6.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
